Added the node package manager/Updated the gitignore to ignore .ds_store/ and node_modules//Added a Dist folder for the generated readme/Added a readme file/Added some pseudo code to the index.js and used the require method to include variables for the packages I need/ Added pseudo code to the generateMarkdown.js
Added the node package manager with npm i, added the packages to ignore in the gitignore file, added a dist folder for an output spot for the generated readme however the readme cannot be generated yet. Added a readme file which will be for the github page. Added some pseudo code to the index.js to outline what I have to do some more, and used the require method to include variables for the packages I need as well. Lastly added pseudo code to the generateMarkdown.js to show what this file will be used for.