Updated README.md to include a table of contents, a credits section, a features section, and changed the visuals section to just be the usage section since now there is a features section.
Updated README.md to include a more acknowledgement, a table of contents, a credits section, a features section, and changed the visuals section to just be the usage section since now there is a features section so its set to a professional standard. Updated index.js added code that works using a variable for the filePath, and then using the file system method of writeFile to select the file path, and the data from the prompt to be placed into the file, and if there is an error throw err then console.log that the README.md has been created. For pushing the prompted questions and the data recieved into the README.md file I used the inquirer package with the .prompt method to take in the questions variable which is globally scoped, and then it will place the answers in the correct spots when the data is generated using a local variable which is local to the init function. This variable will be passed into the prompt, and when it its being passed into the prompt what it does is take the answers to the questions and place them inside the readme.md file that is generated with the data.