Update index.js to have a checkbox for the table of contents so you can select which hyper links you need,  arranged the prompted questions for the variable to be in order of where they go on the README.md that is generated, added a variable to the generation part of the init function so that in the questions section of the readme.md the github username can be added to a link that will be displayed next to the github username, then added an email part to the questions sections so that I can get the email, then below the email you can add instructions on how to reach me with additional questions.
Update index.js to have a checkbox for the table of contents so you can select which hyper links you need for the readme.md, arranged the prompted questions for the const questions variable to be in order of where they go on the README.md that is generated so its readable, added a variable to the generation of the template literal inside the init function so that in the questions section of the readme.md the github username can be added to a link that will be displayed next to the github username, then added an email part to the questions sections so that I can get the email, then below the email you can add instructions on how to reach me with additional questions.  However the readme is indented to the right, and does not provide markup text.