Updated index.js to call the generateMarkdown function, and added the variable for the generateMarkdown function to require the function in the index.js code to use/Updated generateMarkdown.js to include the template literal that will be generated, added the const githubLink variable to allow the github username to be targeted and placed into the template literal so I can display a link in the README.md that is generated, and pseudo coded the three render functions however I have not added the variable for the links so they do not work.
Updated index.js to call the generateMarkdown function, and added the variable for the generateMarkdown function to require the function in the index.js code to use, this way the generateMarkdown function can be utilized. Updated generateMarkdown.js to include the template literal that will be generated, added the const githubLink variable to allow the github username to be targeted and placed into the template literal so I can display a link in the README.md that is generated, and pseudo coded the three render functions however I have not added the variable for the links so they do not work yet.