Update index.js to change the names of the license sections choices/Updated generateMarkdown.js to have the badge links, the license links, and the text for the section however none of these functions work yet they just have variables.
Update index.js to change the names of the license sections choices so the user knows what type of license they are selecting. Updated generateMarkdown.js to have the badge links, the license links, and the text for the section however none of these functions work yet they just have variables.