Updated index.js so that the name of the prompt contribution is now Contributing/Updated the README.md file to have new content so that its more professional/Updated the generateMarkdown.js to include an if statement for the badge links that will loop through and select a badge based on the choice of the license, added an if statement to the renderLicenseLink function to make it so that the parameter license in a conditional if statement is checked based on the selection of the license and then return the correct url for the license if not return nothing, added an if statement to the renderLicenseSection to check it None is selected return "N/A"  if else return a message with the license link if a license is selected when prompted, also changed the generation of the markdown text to have indentation.
Updated index.js so that the name of the prompt contribution is now Contributing so that it follows criteria. Updated the README.md file to have new content so that its more professional that way it can be used later on. Updated the generateMarkdown.js to include an if statement for the renderLicenseBadge links that will loop through and select a badge based on the choice of the license, added an if statement to the renderLicenseLink function to make it so that the parameter license in a conditional if statement is checked based on the selection of the license and then return the correct url for the license if not return nothing, added an if statement to the renderLicenseSection to check it None is selected return "N/A"  if else return a message with the license link if a license is selected when prompted, also changed the generation of the markdown text to have indentation so its more readable.