Update index.js to include an if statement in the writeToFile function so that the dist folder is created and the readme.md for the repo is not overwritten.
Update index.js to include an if statement in the writeToFile function so that the dist folder is created and the readme.md for the repo is not overwritten this way the application runs when the repo is cloned. The if statement on line 111-113 tells the write to file function to check if the dist folder has been created when the application is run from the CLI in VS Code, and if not then create one.