Updated the file structure so that the if statement that creates a folder for the readme.md is no longer needed inside of the writeToFile function/Updated the README.md to include the relevant information needed for the README.md for the repo to be professional.
Updated the file structure so everything except the README.md file is inside of the assets folder. Updated the index.js so that the if statement that creates a folder for the readme.md is no longer needed inside of the writeToFile function. Updated the README.md to include the relevant information needed for the README.md for the repo to be professional.